Updated python client to have a clear function, also added in logic to receive chats on the arduino and display them, added more icons, added more logic, etc.... :) BIG COMMIT, DUNNO WHAT TO PUT!

+++ lib/Global/src/Global.h
@@ -3,7 +3,7 @@
 
 namespace global {
 	constexpr size_t UNIT_NUMBER_BUFF_SIZE = 256;
-	constexpr int BAUD_RATE = 115200;
+	constexpr unsigned long BAUD_RATE = 115200U;
 	extern const char* READY;
 
 	void makeReadyToReceive();
